MagenticUI/sharepoint-prompt-mcp/requirements.txt
mcp==1.27.0
httpx==0.28.1
python-dotenv==1.2.1
uvicorn==0.38.0
